Added "remove_collected" support to trakt_acquired
Hi, I tried to add a remove collected setting to Trakt_acquired plugin. It uses this API function: http://trakt.tv/api-docs/lists-items-delete

The idea is to be able to remove downloaded entry from a specified list. Example: I put "Shrek 8" in my Watchlist, it gets out, Flexget download it and mark it as collected on trakt.tv. With this new function it can also remove it from my Watchlist automatically (and keep my watchlist clean).

I'd like to test it but I don't really know how? Can someone test it for me or explain me how to do so?